[REFACTORING] Remoção de entidade Year Class por motivos de normalização do banco de dados

+++ src/main/java/com/ufpa/lafocabackend/domain/model/dto/input/MemberInputDto.java
@@ -35,7 +35,7 @@ public class MemberInputDto {
 
     private Long functionMemberId;
 
-    private Long yearClassId;
+    private Integer yearClassId;
 
     private List<@Positive @NotNull Long> skillsId;
